Refactoring CLI
The usage of Thor seemed a little off to me. A few commands were Thor
groups even though they didn’t need to be. Some commands were exposing
helper methods as commands.

I also refactored the file locations and namespaces to reflect the CLI
itself a little better (grouping base and generate commands). It should
be a bit easier to reason about now.

I renamed the scaffold command to new (to reflect the actual command
name).
