+++ lib/napa/cli/base/new.rb
@@ -2,31 +2,41 @@ require 'thor'
 require 'active_support/core_ext/string'
 
 module Napa
-  module Generators
-    class ScaffoldGenerator < Thor::Group
+  module CLI
+    class Base < Thor
 
       include Thor::Actions
 
-      source_root "#{File.dirname(__FILE__)}/templates/scaffold"
+      source_root File.expand_path("../../templates/new", __FILE__)
 
-      argument :app_name
-      argument :app_path, optional: true
-      class_option :database, default: 'mysql', aliases: '-d', desc: 'Preconfigure for selected database (options: mysql/postgres/pg)'
+      option :database, default: 'mysql', aliases: '-d', desc: 'Preconfigure for selected database (options: mysql/postgres/pg)'
 
-      def generate
+      desc "new <NAME> [PATH]", "Create a new Napa application"
+      def new(name, path = nil)
         say 'Generating scaffold...'
+
+        @name = name
+        @path = path
 
         @database_gem       = postgres? ? 'pg'         : 'mysql2'
         @database_adapter   = postgres? ? 'postgresql' : 'mysql2'
         @database_encoding  = postgres? ? 'unicode'    : 'utf8'
         @database_user      = postgres? ? ''           : 'root'
 
-        directory ".", (app_path || app_name)
+        directory ".", (path || name)
 
         say 'Done!', :green
       end
 
       no_commands do
+
+        def app_name
+          @name
+        end
+
+        def app_path
+          @path
+        end
 
         def postgres?
           %w[pg postgres].include?(options[:database])
