+++ lib/napa/cli/generate.rb
@@ -1,31 +1,18 @@
-require 'thor'
-require 'napa/generators'
+require "napa/cli/generate/api"
+require "napa/cli/generate/readme"
+require "napa/cli/migration"
 
 module Napa
   module CLI
-
     class Generate < Thor
-      register(
-        Generators::ApiGenerator,
-        'api',
-        'api <api_name>',
-        'Create a Grape API, Model and Entity'
-      )
 
       register(
-        Generators::MigrationGenerator,
+        Migration,
         'migration',
-        'migration <migration_name> [field[:type][:index] field[:type][:index]]',
+        'migration <NAME> [field[:type][:index] field[:type][:index]]',
         'Create a Database Migration'
       )
 
-      register(
-        Generators::ReadmeGenerator,
-        'readme',
-        'readme',
-        'Create a formatted README'
-      )
     end
-
   end
 end
