+++ lib/napa/cli/templates/api/app/apis/%name_tableize%_api.rb.tt
