+++ lib/napa/cli/templates/api/app/models/%name_underscore%.rb.tt
