+++ lib/napa/cli/templates/api/app/representers/%name_underscore%_representer.rb.tt
