+++ lib/napa/cli/templates/api/spec/apis/%name_tableize%_api_spec.rb.tt
