+++ lib/napa/cli/templates/api/spec/models/%name_underscore%_spec.rb.tt
