+++ lib/napa/cli/templates/create_table_migration/%migration_filename%.rb.tt
