+++ lib/napa/cli/templates/migration/%migration_filename%.rb.tt
