+++ lib/napa/cli/templates/new/.env.tt
