+++ lib/napa/cli/templates/new/.gitignore.tt
