+++ lib/napa/cli/templates/new/.rubocop.yml
