+++ lib/napa/cli/templates/new/.ruby-gemset.tt
