+++ lib/napa/cli/templates/new/.ruby-version.tt
