+++ lib/napa/cli/templates/new/Gemfile.tt
