+++ lib/napa/cli/templates/new/README.md
