+++ lib/napa/cli/templates/new/Rakefile
