+++ lib/napa/cli/templates/new/app.rb
