+++ lib/napa/cli/templates/new/app/apis/application_api.rb
