+++ lib/napa/cli/templates/new/app/apis/hello_api.rb.tt
