+++ lib/napa/cli/templates/new/config.ru.tt
