+++ lib/napa/cli/templates/new/config/database.yml.tt
