+++ lib/napa/cli/templates/new/config/initializers/active_record.rb
