+++ lib/napa/cli/templates/new/config/middleware/honeybadger.rb
