+++ lib/napa/cli/templates/new/db/schema.rb
