+++ lib/napa/cli/templates/new/spec/apis/hello_api_spec.rb.tt
