+++ lib/napa/cli/templates/new/spec/docs/readme_spec.rb
