+++ lib/napa/cli/templates/new/spec/spec_helper.rb
