+++ lib/napa/cli/templates/readme/spec/docs/readme_spec.rb
@@ -1,0 +1,7 @@
+require 'spec_helper'
+
+describe 'README' do
+  it 'has been filled out' do
+    pending 'Please fill out the sections that have a :bow:'
+  end
+end
