+++ spec/cli/base/new_spec.rb
@@ -1,8 +1,7 @@
 require 'spec_helper'
-require 'napa/generators/scaffold_generator'
-require 'napa/cli'
+require 'napa/cli/base/new'
 
-describe Napa::Generators::ScaffoldGenerator do
+describe Napa::CLI::Base do
   let(:app_name) { 'my_test_app' }
   let(:app_path) { 'spec/my_different_directory' }
   let(:options) { {} }
