+++ spec/cli/migration_spec.rb
@@ -1,8 +1,8 @@
 require 'spec_helper'
-require 'napa/generators/migration_generator'
-require 'napa/cli'
+require 'napa/cli/migration'
 
-describe Napa::Generators::MigrationGenerator do
+describe Napa::CLI::Migration do
+
   let(:migration_filename) { 'foo' }
   let(:test_migrations_directory) { 'spec/tmp' }
 
@@ -19,7 +19,8 @@ describe Napa::Generators::MigrationGenerator do
 
   describe 'AddFooToBar flew:string:index brew:integer' do
     before do
-      Napa::CLI::Base.new.generate("migration", 'AddFooToBar', 'flew:string:index', 'brew:integer')
+      Napa::CLI::Migration.new(['AddFooToBar', 'flew:string:index', 'brew:integer']).invoke
+
       expected_migration_file = File.join(test_migrations_directory, 'foo.rb')
       @migration_code = File.read(expected_migration_file)
     end
@@ -40,7 +41,7 @@ describe Napa::Generators::MigrationGenerator do
 
   describe 'RemoveFooFromBar flew:string brew:integer' do
     before do
-      Napa::CLI::Base.new.generate("migration", 'RemoveFooFromBar', 'flew:string', 'brew:integer')
+      Napa::CLI::Migration.new(['RemoveFooFromBar', 'flew:string', 'brew:integer']).invoke
       expected_migration_file = File.join(test_migrations_directory, 'foo.rb')
       @migration_code = File.read(expected_migration_file)
     end
@@ -57,7 +58,7 @@ describe Napa::Generators::MigrationGenerator do
 
   describe 'CreateJoinTableFooBar foo bar' do
     before do
-      Napa::CLI::Base.new.generate("migration", 'CreateJoinTableFooBar', 'foo', 'bar')
+      Napa::CLI::Migration.new(['CreateJoinTableFooBar', 'foo', 'bar']).invoke
       expected_migration_file = File.join(test_migrations_directory, 'foo.rb')
       @migration_code = File.read(expected_migration_file)
     end
@@ -78,7 +79,7 @@ describe Napa::Generators::MigrationGenerator do
 
   describe 'CreateSkrillex drops:integer hair:string:index' do
     before do
-      Napa::CLI::Base.new.generate("migration", 'CreateSkrillex', 'drops:integer', 'hair:string:index')
+      Napa::CLI::Migration.new(['CreateSkrillex', 'drops:integer', 'hair:string:index']).invoke
       expected_migration_file = File.join(test_migrations_directory, 'foo.rb')
       @migration_code = File.read(expected_migration_file)
     end
@@ -100,7 +101,5 @@ describe Napa::Generators::MigrationGenerator do
       expect(@migration_code).to match(/add_index :skrillexes, :hair/)
     end
   end
-
-  describe
 
 end
